requirements.txt
certifi=2021.5.30
charset-normalizer=2.0.3
click=8.0.1
colorama=0.4.4
cycler=0.10.0
Flask=2.0.1
idna=3.2
itsdangerous=2.0.1
Jinja2=3.0.1
kiwisolver=1.3.1
MarkupSafe=2.0.1
matplotlib=3.4.2
numpy=1.21.0
pandas=1.3.0
Pillow=8.3.1
pip=20.1.1
pyparsing=2.4.7
python-dateutil=2.8.2
pytz=2021.1
quickchart.io=0.2.0
requests=2.26.0
setuptools=47.1.0
six=1.16.0
urllib3=1.26.6
Werkzeug=2.0.1
